feat(component): add button.jsx and button.stories.js
