+++ src/components/button/Button.stories.js
@@ -1,0 +1,16 @@
+import Button from './Button';
+
+export default {
+	title: 'Dispute_Resolution/Button',
+	component: Button,
+	tags: ['autodocs'],
+	argTypes: {
+		backgroundColor: { control: 'color' },
+	},
+};
+export const Small = {
+	args: {
+		// primary: true,
+		label: 'Button',
+	},
+};
